% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomletter.R
\name{randomletter}
\alias{randomletter}
\title{randomletter(x)}
\usage{
randomletter(x)
}
\arguments{
\item{x}{pretty much does nothing}
}
\value{
a random capital letter from the alphabet
}
\description{
Has a "useless" argument for parallel computation, and returns a random letter
}
